main types to main pages folder
